Remove file duplication
