--- src/gsw/MCT/test/.gitignore
@@ -1,4 +1,0 @@
-openmct
-node_modules
-npm-debug.log
-.idea
