--- src/gsw/MCT/test/example-server/static-server.js
@@ -1,11 +1,0 @@
-var express = require('express');
-
-function StaticServer() {
-    var router = express.Router();
-
-    router.use('/', express.static(__dirname + '/..'));
-
-    return router
-}
-
-module.exports = StaticServer;
