--- src/gsw/MCT/test/historical-telemetry-plugin.js
@@ -1,26 +1,0 @@
-/**
- * Basic historical telemetry plugin.
- */
-
-function HistoricalTelemetryPlugin() {
-    return function install (openmct) {
-        var provider = {
-            supportsRequest: function (domainObject) {
-                return domainObject.type === 'example.telemetry';
-            },
-            request: function (domainObject, options) {
-                var url = '/history/' +
-                    domainObject.identifier.key +
-                    '?start=' + options.start +
-                    '&end=' + options.end;
-
-                return http.get(url)
-                    .then(function (resp) {
-                        return resp.data;
-                    });
-            }
-        };
-
-        openmct.telemetry.addProvider(provider);
-    }
-}
