--- src/gsw/MCT/test/index.html
@@ -1,28 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Open MCT Tutorials</title>
-    <script src="node_modules/openmct/dist/openmct.js"></script>
-    <script src="lib/http.js"></script>
-    <script src="dictionary-plugin.js"></script>
-    <script src="historical-telemetry-plugin.js"></script>
-    <script src="realtime-telemetry-plugin.js"></script>
-</head>
-<body>
-    <script>
-        openmct.setAssetPath('node_modules/openmct/dist');
-        openmct.install(openmct.plugins.LocalStorage());
-        openmct.install(openmct.plugins.MyItems());
-        openmct.install(openmct.plugins.UTCTimeSystem());
-        openmct.time.clock('local', {start: -15 * 60 * 1000, end: 0});
-        openmct.time.timeSystem('utc');
-        openmct.install(openmct.plugins.Espresso());
-
-        openmct.install(DictionaryPlugin());
-        openmct.install(HistoricalTelemetryPlugin());
-        openmct.install(RealtimeTelemetryPlugin());
-
-        openmct.start();
-    </script>
-</body>
-</html>
